+++ src/qtasks/brokers/async_kafka.py
@@ -10,7 +10,7 @@ from uuid import UUID, uuid4
 from time import time
 from typing import TYPE_CHECKING
 
-from qtasks.configs.config_observer import ConfigObserver
+from qtasks.configs.config import QueueConfig
 from qtasks.enums.task_status import TaskStatusEnum
 from qtasks.logs import Logger
 from qtasks.storages.sync_redis import SyncRedisStorage
@@ -94,12 +94,12 @@ class AsyncKafkaBroker(BaseBroker):
                 )
             ] = None,
             config: Annotated[
-                Optional[ConfigObserver],
+                Optional[QueueConfig],
                 Doc(
                     """
-                    Логгер.
+                    Конфиг.
                     
-                    По умолчанию: `qtasks.configs.config_observer.ConfigObserver`.
+                    По умолчанию: `qtasks.configs.config.QueueConfig`.
                     """
                 )
             ] = None