+++ src/qtasks/brokers/base.py
@@ -4,7 +4,6 @@ from uuid import UUID
 from typing_extensions import Annotated, Doc
 from typing import TYPE_CHECKING
 
-from qtasks.configs.config_observer import ConfigObserver
 from qtasks.logs import Logger
 from qtasks.schemas.task import Task
 from qtasks.configs.config import QueueConfig
@@ -65,18 +64,18 @@ class BaseBroker(ABC):
                 )
             ] = None,
             config: Annotated[
-                Optional[ConfigObserver],
+                Optional[QueueConfig],
                 Doc(
                     """
-                    Логгер.
+                    Конфиг.
                     
-                    По умолчанию: `qtasks.configs.ConfigObserver`.
+                    По умолчанию: `qtasks.configs.config.QueueConfig`.
                     """
                 )
             ] = None
         ):
         self.name = name
-        self.config = config or ConfigObserver(QueueConfig())
+        self.config = config or QueueConfig()
         self.storage = storage
         self.log = log.with_subname("Broker") if log else Logger(name=self.name, subname="Broker", default_level=self.config.logs_default_level, format=self.config.logs_format)
         self.plugins: dict[str, "BasePlugin"] = {}
@@ -199,7 +198,7 @@ class BaseBroker(ABC):
 
     def update_config(self,
             config: Annotated[
-                ConfigObserver,
+                QueueConfig,
                 Doc(
                     """
                     Конфиг.