+++ src/qtasks/configs/base.py
@@ -3,7 +3,6 @@ from typing import TYPE_CHECKING, Any, Optional
 from typing_extensions import Annotated, Doc
 
 from qtasks.configs.config import QueueConfig
-from qtasks.configs.config_observer import ConfigObserver
 from qtasks.logs import Logger
 
 if TYPE_CHECKING:
@@ -50,19 +49,19 @@ class BaseGlobalConfig(ABC):
                 )
             ] = None,
             config: Annotated[
-                Optional[ConfigObserver],
+                Optional[QueueConfig],
                 Doc(
                     """
-                    Логгер.
+                    Конфиг.
                     
-                    По умолчанию: `qtasks.configs.config_observer.ConfigObserver`.
+                    По умолчанию: `qtasks.configs.config.QueueConfig`.
                     """
                 )
             ] = None
         ):
         self.name = name
         self.client = None
-        self.config = config or ConfigObserver(QueueConfig())
+        self.config = config or QueueConfig()
         
         self.log = log.with_subname("GlobalConfig") if log else Logger(name=self.name, subname="GlobalConfig", default_level=self.config.logs_default_level, format=self.config.logs_format)
         self.plugins: dict[str, "BasePlugin"] = {}