+++ src/qtasks/starters/base.py
@@ -1,5 +1,4 @@
 from abc import ABC, abstractmethod
-from qtasks.configs.config_observer import ConfigObserver
 from qtasks.logs import Logger
 from typing import TYPE_CHECKING, Optional, Type
 from typing_extensions import Annotated, Doc
@@ -74,18 +73,18 @@ class BaseStarter(ABC):
                 )
             ] = None,
             config: Annotated[
-                Optional[ConfigObserver],
+                Optional[QueueConfig],
                 Doc(
                     """
-                    Логгер.
+                    Конфиг.
                     
-                    По умолчанию: `qtasks.configs.config_observer.ConfigObserver`.
+                    По умолчанию: `qtasks.configs.config.QueueConfig`.
                     """
                 )
             ] = None
         ):
         self.name = name
-        self.config = config or ConfigObserver(QueueConfig())
+        self.config = config or QueueConfig()
         self.log = log.with_subname("Starter") if log else Logger(name=self.name, subname="Starter", default_level=self.config.logs_default_level, format=self.config.logs_format)
         
         self.broker = broker