+++ src/qtasks/starters/sync_starter.py
@@ -1,4 +1,4 @@
-from qtasks.configs.config_observer import ConfigObserver
+from qtasks.configs.config import QueueConfig
 from qtasks.logs import Logger
 from typing import TYPE_CHECKING, Optional
 from typing_extensions import Annotated, Doc
@@ -76,12 +76,12 @@ class SyncStarter(BaseStarter):
                 )
             ] = None,
             config: Annotated[
-                Optional[ConfigObserver],
+                Optional[QueueConfig],
                 Doc(
                     """
-                    Логгер.
+                    Конфиг.
                     
-                    По умолчанию: `qtasks.configs.config_observer.ConfigObserver`.
+                    По умолчанию: `qtasks.configs.config.QueueConfig`.
                     """
                 )
             ] = None