+++ src/qtasks/storages/sync_redis.py
@@ -1,15 +1,13 @@
 import datetime
 import json
 import time
-import traceback
 from typing import Optional, Union
-from httpx import delete
 from typing_extensions import Annotated, Doc
 from uuid import UUID
 import redis
 from typing import TYPE_CHECKING
 
-from qtasks.configs.config_observer import ConfigObserver
+from qtasks.configs.config import QueueConfig
 from qtasks.contrib.redis.sync_queue_client import SyncRedisCommandQueue
 from qtasks.configs.sync_redisglobalconfig import SyncRedisGlobalConfig
 from qtasks.enums.task_status import TaskStatusEnum
@@ -105,12 +103,12 @@ class SyncRedisStorage(BaseStorage):
                 )
             ] = None,
             config: Annotated[
-                Optional[ConfigObserver],
+                Optional[QueueConfig],
                 Doc(
                     """
-                    Логгер.
+                    Конфиг.
                     
-                    По умолчанию: `qtasks.configs.config_observer.ConfigObserver`.
+                    По умолчанию: `qtasks.configs.config.QueueConfig`.
                     """
                 )
             ] = None