+++ src/qtasks/workers/base.py
@@ -1,10 +1,9 @@
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Optional, Type
+from typing import TYPE_CHECKING, Optional
 from uuid import UUID
 from typing_extensions import Annotated, Doc
 
 from qtasks.configs.config import QueueConfig
-from qtasks.configs.config_observer import ConfigObserver
 from qtasks.logs import Logger
 from qtasks.middlewares.task import TaskMiddleware
 from qtasks.schemas.inits import InitsExecSchema
@@ -64,19 +63,19 @@ class BaseWorker(ABC):
                 )
             ] = None,
             config: Annotated[
-                Optional[ConfigObserver],
+                Optional[QueueConfig],
                 Doc(
                     """
-                    Логгер.
+                    Конфиг.
                     
-                    По умолчанию: `qtasks.configs.config_observer.ConfigObserver`.
+                    По умолчанию: `qtasks.configs.config.QueueConfig`.
                     """
                 )
             ] = None
         ):
         self.name = name
         self.broker = broker
-        self.config = config or ConfigObserver(QueueConfig())
+        self.config = config or QueueConfig()
         
         self.log = log.with_subname("Worker") if log else Logger(name=self.name, subname="Worker", default_level=self.config.logs_default_level, format=self.config.logs_format)
         self.init_worker_running: list[InitsExecSchema] = []