+++ docs/ru/changelog.md
@@ -5,7 +5,7 @@
 - Добавлен [`SyncTaskExecutor`](/qtasks/ru/api/executors/sync_task_executor/) и [`AsyncTaskExecutor`](/qtasks/ru/api/executors/async_task_executor/) и добавлен в [`Worker`](/qtasks/ru/api/workers/baseworker/).
 - Добавлены [`SyncRedisCommandQueue`](/qtasks/ru/api/classes/sync_redis_commands/) и [`AsyncRedisCommandQueue`](/qtasks/ru/api/classes/async_redis_commands/).
 - Добавлен [`SyncKafkaBroker`](/qtasks/ru/api/brokers/sync_kafkabroker/).
-- Добавлен [`ConfigObserver`](/qtasks/ru/api/classes/config_observer/) и был заменен в `QueueTasks.config`.
+- Добавлен `ConfigObserver` и был заменен в `QueueTasks.config`.
 - Добавлен [`BaseMiddleware`](/qtasks/ru/api/middlewares/basemiddleware/) и [`TaskMiddleware`](/qtasks/ru/api/middlewares/task_middleware/).
 - Добавлен [`ping()`](/qtasks/ru/api/queuetasks/#qtasks.qtasks.QueueTasks.ping).
 - Добавлен [`flush_all()`](/qtasks/ru/api/queuetasks/#qtasks.qtasks.QueueTasks.flush_all).